refactor: move ArduinoRepository to repository package
