+++ app/src/main/java/org/kabiri/android/usbterminal/data/repository/ArduinoRepository.kt
@@ -1,4 +1,4 @@
-package org.kabiri.android.usbterminal.arduino
+package org.kabiri.android.usbterminal.data.repository
 
 import android.content.Context
 import android.hardware.usb.UsbDevice
@@ -14,6 +14,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import org.kabiri.android.usbterminal.R
+import org.kabiri.android.usbterminal.arduino.ArduinoSerialReceiver
 import org.kabiri.android.usbterminal.domain.IGetCustomBaudRateUseCase
 import org.kabiri.android.usbterminal.model.defaultBaudRate
 import javax.inject.Inject
